+++ cmd/run.go
@@ -31,14 +31,14 @@ import (
 	"github.com/openfga/openfga/internal/server"
 	serverErrors "github.com/openfga/openfga/internal/server/errors"
 	"github.com/openfga/openfga/internal/server/health"
-	"github.com/openfga/openfga/internal/storage"
-	"github.com/openfga/openfga/internal/storage/caching"
-	"github.com/openfga/openfga/internal/storage/common"
-	"github.com/openfga/openfga/internal/storage/memory"
-	"github.com/openfga/openfga/internal/storage/mysql"
-	"github.com/openfga/openfga/internal/storage/postgres"
 	"github.com/openfga/openfga/pkg/encoder"
 	"github.com/openfga/openfga/pkg/logger"
+	"github.com/openfga/openfga/pkg/storage"
+	"github.com/openfga/openfga/pkg/storage/caching"
+	"github.com/openfga/openfga/pkg/storage/common"
+	"github.com/openfga/openfga/pkg/storage/memory"
+	"github.com/openfga/openfga/pkg/storage/mysql"
+	"github.com/openfga/openfga/pkg/storage/postgres"
 	"github.com/openfga/openfga/pkg/telemetry"
 	"github.com/rs/cors"
 	"github.com/spf13/cobra"
