+++ internal/server/server_test.go
@@ -14,13 +14,13 @@ import (
 	"github.com/openfga/openfga/internal/gateway"
 	serverErrors "github.com/openfga/openfga/internal/server/errors"
 	"github.com/openfga/openfga/internal/server/test"
-	"github.com/openfga/openfga/internal/storage"
-	"github.com/openfga/openfga/internal/storage/common"
-	"github.com/openfga/openfga/internal/storage/memory"
-	mockstorage "github.com/openfga/openfga/internal/storage/mocks"
-	"github.com/openfga/openfga/internal/storage/mysql"
-	"github.com/openfga/openfga/internal/storage/postgres"
 	"github.com/openfga/openfga/pkg/logger"
+	"github.com/openfga/openfga/pkg/storage"
+	"github.com/openfga/openfga/pkg/storage/common"
+	"github.com/openfga/openfga/pkg/storage/memory"
+	mockstorage "github.com/openfga/openfga/pkg/storage/mocks"
+	"github.com/openfga/openfga/pkg/storage/mysql"
+	"github.com/openfga/openfga/pkg/storage/postgres"
 	"github.com/openfga/openfga/pkg/telemetry"
 	storagefixtures "github.com/openfga/openfga/pkg/testfixtures/storage"
 	"github.com/openfga/openfga/pkg/typesystem"
