+++ internal/server/test/delete_store.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	commands2 "github.com/openfga/openfga/internal/server/commands"
-	"github.com/openfga/openfga/internal/storage"
 	"github.com/openfga/openfga/pkg/logger"
+	"github.com/openfga/openfga/pkg/storage"
 	openfgapb "go.buf.build/openfga/go/openfga/api/openfga/v1"
 )
