+++ internal/server/test/read_assertions.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/openfga/openfga/internal/server/commands"
-	"github.com/openfga/openfga/internal/storage"
 	"github.com/openfga/openfga/pkg/logger"
+	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/testutils"
 	openfgapb "go.buf.build/openfga/go/openfga/api/openfga/v1"
 )
