+++ pkg/storage/caching/caching.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/karlseguin/ccache/v2"
-	"github.com/openfga/openfga/internal/storage"
+	"github.com/openfga/openfga/pkg/storage"
 	openfgapb "go.buf.build/openfga/go/openfga/api/openfga/v1"
 )
